+++ src/test/java/com/braintreegateway/unittest/AndroidPayNetworkTokenOptionsTest.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 
 import org.custommonkey.xmlunit.XMLAssert;
 import org.custommonkey.xmlunit.XMLUnit;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.xml.sax.SAXException;
 
 import com.braintreegateway.AndroidPayNetworkTokenOptionsRequest;