+++ src/test/java/com/braintreegateway/unittest/MetaCheckoutCardDetailsTest.java
@@ -1,11 +1,9 @@
 package com.braintreegateway.unittest;
 
-import static org.junit.Assert.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
-import com.braintreegateway.MetaCheckoutCard;
 import com.braintreegateway.MetaCheckoutCardDetails;
 import com.braintreegateway.util.SimpleNodeWrapper;