+++ src/test/java/com/braintreegateway/unittest/MetaCheckoutCardTest.java
@@ -1,6 +1,6 @@
 package com.braintreegateway.unittest;
 
-import static org.junit.Assert.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;